Move CoordinatedShutdown support to internal package
